scraper_service/requirements.txt
requests==2.31.0
cloudscraper>=1.2.71
beautifulsoup4==4.12.3
supabase>=2.11.0
schedule==1.2.1
python-dotenv==1.0.1
pytest==8.0.0
websockets>=13.0
Pillow
boto3>=1.34.0
rarfile>=4.0
py7zr>=0.20.0
